requirements.txt
flask
flask-cors
google-cloud-texttospeech
gunicorn
openai